Add Day 5 solution and tests
Initial solution. Performance on Part 2 is terrible!

I tried a binary search to improve the performance, but the method produced the wrong answer - it appears that the location ranges are non-continuous in mapping back to the seed ranges.

I could try forward mapping the seed range edges to try and detect the ranges; however, with all the intermediate mapping layers this also is not guaranteed to produce contiguous location ranges.